% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{coastlines}
\alias{coastline10}
\alias{coastlines}
\alias{coastlines10}
\title{world coastlines from Natural Earth}
\format{A \code{SpatialLinesDataFrame}}
\source{
\url{http//www.naturalearthdata.com/download/10m/physical/ne_10m_coastline.zip}
}
\usage{
coastline10
}
\description{
coastline lines at scale 1:10m (large).
}
\keyword{datasets}

