% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{map_units}
\alias{map_units}
\alias{map_units10}
\title{world map_unit polygons from Natural Earth}
\format{
A \code{SpatialPolygonsDataFrame}

An object of class \code{sf} (inherits from \code{data.frame}) with 298 rows and 169 columns.
}
\source{
\url{http//www.naturalearthdata.com/download/10m/cultural/ne_10m_admin_0_map_units.zip}
}
\usage{
map_units10
}
\description{
map_unit polygons at scale 1:10m (large).
There are more map_units than countries e.g. United Kingdom is split into England, Scotland, Wales and Northern Ireland.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame with attributes}
}}

\keyword{datasets}
