% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{sovereignty}
\alias{sovereignty}
\alias{sovereignty10}
\title{world sovereignty polygons from Natural Earth}
\format{
A \code{SpatialPolygonsDataFrame}

An object of class \code{sf} (inherits from \code{data.frame}) with 209 rows and 169 columns.
}
\source{
\url{http//www.naturalearthdata.com/download/10m/cultural/ne_10m_admin_0_sovereignty.zip}
}
\usage{
sovereignty10
}
\description{
sovereignty polygons at scale 1:10m (large).
There are fewer sovereign states than countries e.g. the Falkland Islands are included with the United Kingdom.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame with attributes}
}}

\keyword{datasets}
