% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{states}
\alias{states}
\alias{states10}
\title{state (admin level 1) polygons from Natural Earth}
\format{
A \code{SpatialPolygonsDataFrame}

An object of class \code{sf} (inherits from \code{data.frame}) with 4596 rows and 122 columns.
}
\source{
\url{http//www.naturalearthdata.com/download/10m/cultural/ne_10m_admin_1_states.zip}
}
\usage{
states10
}
\description{
For all world countries at scale 1:10m (high).
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame with attributes}
}}

\keyword{datasets}
