% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{countries}
\alias{countries}
\alias{countries10}
\title{world country polygons from Natural Earth}
\format{
A \code{sf} \code{data.frame}

An object of class \code{sf} (inherits from \code{data.frame}) with 258 rows and 169 columns.
}
\source{
\url{http//www.naturalearthdata.com/download/10m/cultural/ne_10m_admin_0_countries.zip}
}
\usage{
countries10
}
\description{
at scale 1:10m (large)
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame with country attributes}
}}

\keyword{datasets}
