% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaturalearthhires-package.r
\docType{package}
\name{rnaturalearthhires}
\alias{rnaturalearthhires}
\title{rnaturalearthhires : high resolution world vector maps from Natural Earth}
\description{
Facilitates world mapping by making \href{http://www.naturalearthdata.com/}{Natural Earth} map data more easily available to R users.
}
\seealso{
\code{\link[rnaturalearth]{ne_countries}} \code{\link[rnaturalearth]{ne_states}} \code{\link[rnaturalearth]{ne_download}}
}
